functions code tidying
