+++ functions/src/DaveHakkensNL/avatarMigrate.ts
@@ -8,14 +8,14 @@ import { downloadFileToTemp, IFileWriteMeta } from '../Utils/file.utils'
 export const migrateAvatar = async (imgUrl: string, userId: string) => {
   // download
   const localImgMeta = (await downloadFileToTemp(imgUrl)) as IFileWriteMeta
-  // convert
+  // convert and upload
+  // note, will operate sequentially on the same file
   const avatar = await resizeLocalImage(localImgMeta.filePath, 'avatar', 'jpg')
-  const thumb = await resizeLocalImage(localImgMeta.filePath, 'thumb', 'jpg')
-  // upload
   await uploadLocalFileToStorage(avatar, `avatars/${userId}.jpg`, {})
+  const thumb = await resizeLocalImage(localImgMeta.filePath, 'thumb', 'jpg')
   await uploadLocalFileToStorage(thumb, `avatars/${userId}.thumb.jpg`, {})
-  console.log('file uploaded')
+  // get meta and return
   const avatarMeta = await getStorageFileMeta(`avatars/${userId}.jpg`)
-  const thumbMeta = await getStorageFileMeta(`avatars/${userId}.jpg`)
+  const thumbMeta = await getStorageFileMeta(`avatars/${userId}.thumb.jpg`)
   return { avatar: avatarMeta, thumb: thumbMeta }
 }
