+++ functions/src/DaveHakkensNL/index.ts
@@ -1,0 +1,8 @@
+import { migrateAvatar } from './avatarMigrate'
+import { updateDHUserIds, getDHUserProfile } from './dataMigrate'
+
+export default {
+  updateDHUserIds,
+  migrateAvatar,
+  getDHUserProfile,
+}
